SwasthyaSetu-main (3)/SwasthyaSetu-main/ayushWhoSearch/requirements.txt
# Core dependencies
requests>=2.31.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil
rapidfuzz
# Machine Learning and NLP dependencies
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
psycopg2-binary
sqlalchemy
# Standard library modules (included with Python)
# csv - built-in
# json - built-in
# os - built-in
# uuid - built-in
# datetime - built-in
torch
transformers
sentence-transformers
googletrans==4.0.0-rc1
pytest
